pyproject.toml
[tool.poetry]
name = "junos-rest"
version = "0.0.1"
description = "Functions to interact with Juniper JunOS devices via REST API"
authors = ["Matt Love <matt@allroads.io>"]
license = "WTFPL"

[tool.poetry.dependencies]
python = "^3.6"
httpx = "^0.11"
xmltodict = "^0.12.0"
ujson = "^1.35"
boltons = "^20.0.0"
pydantic = "^1.4"
pyyaml = "^5.3"
loguru = "^0.4.1"
click = "^7.0"
pygments = "^2.5.2"
uvloop = "^0.14.0"
stackprinter = "^0.2.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
bandit = "^1.6.2"
flake8 = "^3.7.9"
flake8-bandit = "^2.1.2"
flake8-black = "^0.1.1"
flake8-breakpoint = "^1.1.0"
flake8-bugbear = "^20.1.0"
flake8-builtins = "^1.4.2"
flake8-comprehensions = "^3.1.4"
flake8-deprecated = "^1.3"
flake8-eradicate = "^0.2.4"
flake8-if-expr = "^1.0.0"
flake8-isort = "^2.8.0"
flake8-pie = "^0.4.2"
flake8-plugin-utils = "^1.0.0"
flake8-polyfill = "^1.0.2"
flake8-print = "^3.1.4"
flake8-return = "^1.1.1"
pep8-naming = "^0.9.1"
flake8-docstrings = "^1.5.0"
mccabe = "^0.6.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
